Refactor routes for the feature. Commented out unfinished code
